+++ projects/common/lib/models/person.model.ts
@@ -1,7 +1,6 @@
-import * as moment_ from 'moment';
+
 import { Base } from './base';
-import { SimpleDate } from './simple-date.interface';
-const moment = moment_;
+import { format, startOfToday, differenceInYears } from 'date-fns';
 
 /**
  * Person, each project can extend this person class
@@ -16,37 +15,20 @@ export class Person extends Base {
   public lastName: string;
 
   /** Format to display birthdate */
-  public dobFormat = 'YYYY/MM/DD';
+  public dobFormat = 'yyyy/MM/dd';
+  public apiDobFormat = 'yyyyMMdd';
 
-  // Initialize dob to nulls
-  public dateOfBirth: SimpleDate = { year: null, month: null, day: null };
+  // Initialize dob to nulls - To be removed
+  public dateOfBirth: Date;
 
   /** Returns DoB in YYYYMMDD format, used by API. */
   get dateOfBirthShort(): string {
-    return this.isDobEmpty() ? null :
-        moment( {
-          year: this.dateOfBirth.year,
-          month: this.dateOfBirth.month - 1,
-          day: this.dateOfBirth.day
-        }).format( 'YYYYMMDD' );
+    return this.dateOfBirth ? format( this.dateOfBirth, this.apiDobFormat ) : null;
   }
 
   /** Returns DoB in dobFormat (default: YYYY/MM/DD), for display purposes */
   get formatDateOfBirth(): string {
-    return this.isDobEmpty() ? null :
-        moment( {
-          year: this.dateOfBirth.year,
-          month: this.dateOfBirth.month - 1,
-          day: this.dateOfBirth.day
-        }).format( this.dobFormat );
-  }
-
-  /** Indicates whether or not the date of birth is empty */
-  isDobEmpty(): boolean {
-    return Object.keys( this.dateOfBirth )
-        .map( key => this.dateOfBirth[key] )
-        .filter( x => x ) // Filter out null/undefined
-        .length !== 3;
+    return this.dateOfBirth ? format( this.dateOfBirth, this.dobFormat ) : null;
   }
 
   /** Concatenates the first and last name together */
@@ -81,10 +63,28 @@ export class Person extends Base {
     }
   }
 
+  /** Returns the person's full name - first middle last name concatenated together */
+  get fullname() {
+    let _name = null;
+
+    if ( this.firstName ) {
+      _name = this.firstName;
+    }
+
+    if ( this.middleName ) {
+      _name = _name ? _name.concat( ' ' + this.middleName ) : this.middleName;
+    }
+
+    if ( this.lastName ) {
+      _name = _name ? _name.concat( ' ' + this.lastName ) : this.lastName;
+    }
+
+    return _name;
+  }
+
   /** Calculates the age from date of birth */
   getAge(): Number {
-    const dobDt = new Date( this.dateOfBirth.year, this.dateOfBirth.month, this.dateOfBirth.day );
-    return Math.ceil( moment( ).diff( dobDt, 'year', true ) );
+    return differenceInYears( this.dateOfBirth, startOfToday() );
   }
 
   /* Copy function */
@@ -92,8 +92,6 @@ export class Person extends Base {
     this.firstName = object.firstName;
     this.middleName = object.middleName;
     this.lastName = object.lastName;
-    this.dateOfBirth.month = object.dateOfBirth.month;
-    this.dateOfBirth.day = object.dateOfBirth.day;
-    this.dateOfBirth.year = object.dateOfBirth.year;
+    this.dateOfBirth = object.dateOfBirth;
   }
 }